rename library
