+++ lib/design-tokens.js
@@ -1,5 +1,5 @@
-const StyleDictionary = require('style-dictionary');
-const fs = require("fs")
+const StyleDictionary = require("style-dictionary");
+const fs = require("fs");
 
 function createDictionary(root, filename) {
   const name = filename.split(".")[0];
@@ -8,22 +8,26 @@ function createDictionary(root, filename) {
     platforms: {
       scss: {
         transformGroup: "scss",
-        buildPath: `${root}/scss/`, 
-        files: [{
-          destination: `${name}.scss`,
-          format: "scss/variables"
-        }]
+        buildPath: `${root}/scss/`,
+        files: [
+          {
+            destination: `${name}.scss`,
+            format: "scss/variables",
+          },
+        ],
       },
       css: {
         transformGroup: "css",
-        buildPath: `${root}/css/`, 
-        files: [{
-          destination: `${name}.css`,
-          format: "css/variables"
-        }]
-      }
+        buildPath: `${root}/css/`,
+        files: [
+          {
+            destination: `${name}.css`,
+            format: "css/variables",
+          },
+        ],
+      },
     },
-  })
+  });
 }
 
 function getFiles() {
@@ -34,14 +38,14 @@ function generate(root) {
   try {
     const files = getFiles();
     for (const file of files) {
-      const config = createDictionary(root, file)
+      const config = createDictionary(root, file);
       config.buildAllPlatforms();
     }
-  } catch(e) {
+  } catch (e) {
     console.error("ERROR", e.message);
   }
 }
 
 module.exports = {
-  generate: generate
-}
+  generate: generate,
+};
