+++ package.json
@@ -1,10 +1,10 @@
 {
-  "name": "@abgov/style-dictionary",
+  "name": "@abgov/design-tokens",
   "version": "0.0.0-development",
   "main": "./index.js",
   "repository": {
     "type": "git",
-    "url": "https://github.com/GovAlta/style-dictionary.git"
+    "url": "https://github.com/GovAlta/design-tokens.git"
   },
   "keywords": [
     "GoA",
@@ -12,7 +12,7 @@
     "DDI",
     "style",
     "dictionary",
-    "style-dictionary"
+    "design-tokens"
   ],
   "devDependencies": {
     "jest": "^29.2.1",
@@ -28,9 +28,11 @@
     "semantic-release": "semantic-release"
   },
   "bugs": {
-    "url": "https://github.com/GovAlta/style-dictionary/issues"
+    "url": "https://github.com/GovAlta/design-tokens/issues"
   },
   "release": {
-    "branches": ["main"]
+    "branches": [
+      "main"
+    ]
   }
 }
